requirements.txt
# Core ML framework
torch>=2.1.0

# Unsloth — fast QLoRA fine-tuning for consumer GPUs
unsloth[colab-new]>=2024.8

# HuggingFace ecosystem
transformers>=4.44.0
datasets>=2.20.0
peft>=0.12.0
trl>=0.9.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
huggingface_hub>=0.24.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
